integrated_implementation/requirements.txt
torch>=2.0.0
triton>=2.0.0
transformers>=4.30.0
numpy>=1.21.0,<2.0.0  # Pin to avoid compatibility issues
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.9.0  # For statistical analysis
pandas>=1.5.0  # For data handling
scikit-learn>=1.2.0  # For additional metrics